[BE] refactor(playinghistory): config 패키지 위치 수정 #808
