+++ backend/app/src/main/java/com/onair/hearit/app/playinghistory/infrastructure/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.onair.hearit.app.playinghistory.infrastructure.buffer;
+package com.onair.hearit.app.playinghistory.infrastructure.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
